BACKLOG-9984 / Handle synchronization of factory configurations : - Added new key to identify configurations based on fileinstall filename when available. - Improved lookup of matching local configuration by using filename in addition to single PID (findLocalConfiguration) - Correctly create factoryConfig when needed (createLocalConfiguration). - Updated command to gather configs with same filename but different PIDs - Avoid loop when syncing factoryPid without filename (canDistributeConfig)

+++ config/src/main/java/org/apache/karaf/cellar/config/ConfigurationSynchronizer.java
@@ -29,10 +29,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.Dictionary;
-import java.util.Map;
-import java.util.Properties;
-import java.util.Set;
+import java.util.*;
 
 /**
  * The ConfigurationSynchronizer is called when Cellar starts or when a node joins a cluster group.
@@ -130,16 +127,21 @@ public class ConfigurationSynchronizer extends ConfigurationSupport implements S
                         Dictionary clusterDictionary = clusterConfigurations.get(pid);
                         try {
                             // update the local configuration if needed
-                            Configuration localConfiguration = configurationAdmin.getConfiguration(pid, null);
+                            Configuration localConfiguration = findLocalConfiguration(pid, clusterDictionary);
+                            if (localConfiguration == null) {
+                                // Create new configuration
+                                localConfiguration = createLocalConfiguration(pid, clusterDictionary);
+                            }
                             Dictionary localDictionary = localConfiguration.getProperties();
                             if (localDictionary == null)
                                 localDictionary = new Properties();
 
                             localDictionary = filter(localDictionary);
-                            if (!equals(clusterDictionary, localDictionary)) {
+                            if (!equals(clusterDictionary, localDictionary) && canDistributeConfig(localDictionary)) {
                                 LOGGER.debug("CELLAR CONFIG: updating configration {} on node", pid);
+                                clusterDictionary = convertPropertiesFromCluster(clusterDictionary);
                                 localConfiguration.update((Dictionary) clusterDictionary);
-                                persistConfiguration(configurationAdmin, pid, clusterDictionary);
+                                persistConfiguration(localConfiguration);
                             }
                         } catch (IOException ex) {
                             LOGGER.error("CELLAR CONFIG: failed to read local configuration", ex);
@@ -149,9 +151,14 @@ public class ConfigurationSynchronizer extends ConfigurationSupport implements S
                 // cleanup the local configurations not present on the cluster if the node is not the first one in the cluster
                 if (CellarUtils.doCleanupResourcesNotPresentInCluster(configurationAdmin) && getSynchronizerMap().containsKey(Constants.CONFIGURATION_MAP + Configurations.SEPARATOR + groupName)) {
                     try {
+                        Set<String> filenames = new HashSet();
+                        for (Properties configuration : clusterConfigurations.values()) {
+                            filenames.add(getKarafFilename(configuration));
+                        }
+                        filenames.remove(null);
                         for (Configuration configuration : configurationAdmin.listConfigurations(null)) {
                             String pid = configuration.getPid();
-                            if (!clusterConfigurations.containsKey(pid) && isAllowed(group, Constants.CATEGORY, pid, EventType.INBOUND)) {
+                            if (!clusterConfigurations.containsKey(pid) && !filenames.contains(getKarafFilename(configuration.getProperties())) && isAllowed(group, Constants.CATEGORY, pid, EventType.INBOUND)) {
                                 LOGGER.debug("CELLAR CONFIG: deleting local configuration {} which is not present in cluster", pid);
                                 configuration.delete();
                             }
@@ -160,6 +167,8 @@ public class ConfigurationSynchronizer extends ConfigurationSupport implements S
                         LOGGER.warn("Can't get local configurations", e);
                     }
                 }
+            } catch (Exception ex) {
+                LOGGER.error("CELLAR CONFIG: failed to read cluster configuration", ex);
             } finally {
                 Thread.currentThread().setContextClassLoader(originalClassLoader);
             }
@@ -208,7 +217,7 @@ public class ConfigurationSynchronizer extends ConfigurationSupport implements S
                                 eventProducer.produce(event);
                             } else {
                                 Dictionary clusterDictionary = clusterConfigurations.get(pid);
-                                if (!equals(clusterDictionary, localDictionary)) {
+                                if (!equals(clusterDictionary, localDictionary) && canDistributeConfig(localDictionary)) {
                                     LOGGER.debug("CELLAR CONFIG: updating configuration pid {} on the cluster", pid);
                                     // update cluster configurations
                                     clusterConfigurations.put(pid, dictionaryToProperties(localDictionary));
